requirements.txt
# Web + schema
fastapi~=0.116.1
uvicorn[standard]>=0.30,<0.35
pydantic~=2.11.7
python-multipart>=0.0.9,<1

# DB
SQLAlchemy>=2.0,<2.1
psycopg[binary]>=3.1,<3.2  # PostgreSQL driver

# Storage / cache
boto3>=1.34,<2.0
redis>=5.0,<6.0

# Vision / ML
numpy>=1.26,<3.0
opencv-python-headless>=4.9,<4.11
Pillow>=10.3,<11.0
scikit-learn>=1.4,<1.6
insightface>=0.7,<0.9
onnxruntime>=1.17,<2.0   # use onnxruntime-gpu if you have CUDA

pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0

gunicorn>=21,<22

# Optional perf
# orjson>=3.9,<4.0
